+++ src/game_board.h
@@ -1,6 +1,5 @@
-//
-// Created by home on 15.02.2022.
-//
+// This is a personal academic project. Dear PVS-Studio, please check it.
+// PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
 #ifndef TEMPLATE_GAME_BOARD_H
 #define TEMPLATE_GAME_BOARD_H
@@ -21,7 +20,7 @@ public:
 
     ErrorCode fil_cell(size_t i, size_t j, int state);
 
-    ErrorCode valid_move(size_t i, size_t j);
+    ErrorCode valid_move(int i, int j);
 
     bool is_filled();
